requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cssselect==1.1.0
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
lxml==4.6.3
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.1
readability-lxml==0.8.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
uritemplate==4.1.1
urllib3==1.26.7
